More revert of unused changes in fluvio bin
