+++ crates/fluvio-channel/src/lib.rs
@@ -4,9 +4,6 @@ use std::fs::{File, create_dir_all, read_to_string};
 use std::io::{ErrorKind, Error as IoError, Write};
 use color_eyre::Result;
 use color_eyre::eyre::eyre;
-//use crate::{Result, CliError};
-//use fluvio_cli::install::{install_bin, fetch_latest_version, fetch_package_file, install_println};
-//use fluvio_index::{PackageId, HttpAgent};
 use fluvio::FluvioError;
 use fluvio::config::ConfigError;
 use fluvio_types::defaults::CLI_CONFIG_PATH;
