+++ crates/fluvio-cli/Cargo.toml
@@ -67,8 +67,6 @@ crossterm = { version = "0.22.1", features = ['event-stream']}
 tui = { version = "0.16.0", default-features = false, features = ['crossterm'] }
 futures = "0.3"
 humantime = "2.1.0"
-toml = "0.5.5"
-cfg-if = "1.0"
 
 # Fluvio dependencies
 k8-config = { version = "1.4.0", optional = true }
