+++ crates/fluvio-cli/src/topic/create.rs
@@ -158,7 +158,7 @@ impl CreateTopicOpt {
     }
 }
 
-#[derive(Debug, StructOpt, Clone)]
+#[derive(Debug, StructOpt)]
 pub struct TopicConfigOpt {
     /// Retention time (round to seconds)
     /// Ex: '1h', '2d 10s', '7 days' (default)
